Simulating

Simulation73.AbstractModel — Type

A AbstractModel specifies all parameters of a system.

Simulation73.Execution — Type

An Execution holds a Simulation and the solution obtained by running the Simulation.

Simulation73.Simulation — Type

A Simulation holds an AbstractModel to be solved, the space on which to solve it, the time for which to solve it, the initial value, and various solver options.

Simulation73.make_system_mutator — Method

make_system_mutator(simulation)

Construct the differential function to be provided to the ODE solver.