Variables

Simulation73.DeconstructorFixingVariables — Type

deconstruct(..., DeconstructorFixingVariables())

Deconstructs as normal (i.e. Deconstructor) except on AbstractVariable parameters, which are fixed to their default value.

Simulation73.UnboundedVariable — Type

A parameter to vary, without bounds.

Simulation73.deconstruct — Method

deconstruct(ad::AbstractDeconstructor, param)

Deconstruct a parameter into nested tuples of the form (type,value).

    Parameter{Varying{T}} |> deconstruct |> reconstruct # isa Parameter{Varying{T}}

Simulation73.variables — Method

Initialize all variables to their default value.

Simulation73.variables — Method

Takes model with variable parameters, and returns default variable values and indices to those variables.