Public Documentation

Documentation for Simulation73's public interface.

See Internal Documentation for internal package docs covering all private functions.

Index

Simulation73.AbstractSpace

Simulation73.abs_difference

Simulation73.abs_difference_periodic

Simulation73.coordinates

Simulation73.differences

Simulation73.discrete_segment

Simulation73.population_timepoint

Public Interface

Simulation73.AbstractSpace — Type

AbstractSpace{T,N_ARR,N_CDT} with distance-type T, storable in array of dimension N_ARR, and with point-coordinates of dimension N_CDT.

Simulation73.abs_difference — Method

abs_difference(edge)

Return the difference between two points in euclidean space, given an edge between those points.

Example

julia> abs_difference( (5,1) )
4

julia> abs_difference( ((2,2), (5,-5)) )
(3, 7)

Simulation73.abs_difference_periodic — Method

abs_difference_periodic(edge, period)

Return the difference between two points in euclidean space as in abs_difference, but let the space wrap with period.

Example

julia> abs_difference_periodic( (5,1), 3 )
-1

julia> abs_difference_periodic( ((2,2), (5,-5)), (3,4) )
(0, -3)

Simulation73.coordinates — Method

coordinates(space::AbstractSpace)

Return an object in the shape of the space where each element is the coordinate of that element.

Simulation73.differences — Method

distances(space)

Return the distances between every pair of points in space

Simulation73.discrete_segment — Method

discrete_segment(extent, n_points)

Return an object containing n_points equidistant coordinates of a segment of length extent centered at 0. If you want 0 to be an element of the segment, make sure n_points is odd.

Example

julia> seg = discrete_segment(0.0, 5.0, 7);

julia> length(seg) == 7
true

julia> seg[end] - seg[1] ≈ 5.0
true

Simulation73.population_timepoint — Method

population_timepoint(solution, pop_dx, time_dx)

Return spatial frame for a given population pop_dx and time time_dx.